Incorporate GitHub Actions and resolve `brms` dependency
Added .yml file to use GitHub Actions for automated docker image builds, and pushing to DockerHub upon a tagged commit. Also resolved the new dependency introduced 07/18/2020 for the `brms` package, the `cmdstanr` package that is not found on CRAN. Needed to add an extra argument to `--repos` for the `install2.r` script.